Write a function that checks for uppercase character.
Write a function that checks for a digit (0 through 9).

Write a function that multiplies two integers.

Write a function that prints the numbers, from 0 to 9, followed by a new line.

Write a function that prints the numbers, from 0 to 9, followed by a new line.

Write a function that prints 10 times the numbers, from 0 to 14, followed by a new line.

Write a function that draws a straight line in the terminal.

7.Write a function that draws a diagonal line on the terminal.

Write a function that prints a square, followed by a new line.

Write a program that prints the numbers from 1 to 100, followed by a new line.

Write a function that prints a triangle, followed by a new line.

Write a program that finds and prints the largest prime factor of the number 612852475143, followed by a new line.

clear
clear

ls